requirements.txt
python-telegram-bot>=21.3
openai>=1.35.0
python-dotenv>=1.0.1
